% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ARI}
\alias{ARI}
\title{calculate the value of adjusted Rand Index}
\usage{
ARI(k, n, true_label, label)
}
\arguments{
\item{k}{number of cluster}

\item{n}{nodes number}

\item{true_label}{the true label of nodes n x 1 vector}

\item{label}{the generated label of nodes n x 1 vector}
}
\value{
the value of adjusted Rand Index
}
\description{
calculate the value of adjusted Rand Index
}
