% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different_model.R
\name{DDSYM}
\alias{DDSYM}
\title{use DDSYM model to cluster the directed graph}
\usage{
DDSYM(k, g1, n)
}
\arguments{
\item{k}{number of clusters}

\item{g1}{Adjacency Matrix  generated from function DSBM}

\item{n}{nodes number}
}
\value{
kmeans result
}
\description{
use DDSYM model to cluster the directed graph
}
\examples{
 G<- DSBM(4,100,0.8,0.4)
 DDSYM(4,G$g1,100)

}
