% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different_model.R
\name{different_model}
\alias{different_model}
\title{use different model to cluster the graph}
\usage{
different_model(k, cg1, cg2, g1, D, n)
}
\arguments{
\item{k}{number of clusters}

\item{cg1}{complex matrix 1 generated from function DSBM}

\item{cg2}{complex matrix 2 generated from function DSBM}

\item{g1}{symmetric matrix generated from function DSBM}

\item{D}{diagonal matrix whose parameter is the degree of nodes}

\item{n}{nodes number}
}
\value{
list(vectors1,kmeansresult1,kmeansresult2)
}
\description{
use different model to cluster the graph
}
\examples{
 G<- DSBM(4,100,0.8,0.4)
different_model(4,G$cg1,G$cg2,G$g1,G$D,100)

}
