% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDSBM.R
\name{DSBM}
\alias{DSBM}
\title{Generate Directed SBM}
\usage{
DSBM(k, n, p, q)
}
\arguments{
\item{k}{number of cluster s}

\item{n}{nodes number}

\item{p}{probability of connecting nodes in the same cluster}

\item{q}{probability of connecting nodes in the different cluster}
}
\value{
list(symmetric matrix,complex matrix 1 ,complex matrix 2,diag(x))
}
\description{
Generate Directed SBM
}
\examples{
DSBM(4,100,0.8,0.4)
}
