\name{SA24204166-package}
\alias{SA24204166-package}
\alias{SA24204166}
\docType{package}
\title{
  Directed SBM functions
}
\description{
 generate DSBM and use DDSYM ,HERM ,SimpleHERM to deal with it.
}
\details{
  function DSBM is used to generate a directed SBM ,different_model use three different method to deal with it,and function ARI compute the value that
  is in [0,1] which indicates the differences between true label and label generated.
}
\author{
SA24204166xsy

Maintainer: SA24204166 <ustcxsy@mail.ustc.edu.cn>
}

\keyword{ package }

