% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different_model.R
\name{Simpleherm}
\alias{Simpleherm}
\title{use Simpleherm model to cluster the graph}
\usage{
Simpleherm(k, cg2, D, n)
}
\arguments{
\item{k}{number of clusters}

\item{cg2}{complex matrix 2 generated from function DSBM}

\item{D}{diagonal matrix whose parameter is the degree of nodes}

\item{n}{nodes number}
}
\value{
list(vectors1,kmeansresult1,kmeansresult2)
}
\description{
use Simpleherm model to cluster the graph
}
\examples{
 G<- DSBM(4,100,0.8,0.4)
 Simpleherm(4,G$cg2,G$D,100)

}
