% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ARI}
\alias{ARI}
\title{calculate the value of adjusted Rand Index}
\usage{
ARI(k, n, true_label, label)
}
\arguments{
\item{k}{number of cluster}

\item{n}{nodes number}

\item{true_label}{the true label of nodes n x 1 vector}

\item{label}{the generated label of nodes n x 1 vector}
}
\value{
the value of adjusted Rand Index
}
\description{
calculate the value of adjusted Rand Index
}
\examples{
G=DSBM(4,100,0.8,0.4)
H=different_model(4,G$cg1,G$cg2,G$g1,G$D,100)
ARI(4,100,rep(1:4,each= 25),(H$km1$cluster))

}
